refactor: move db init to module
